requirements_dev.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.10.0
astroid==4.0.2
bandit==1.8.6
black==25.11.0
click==8.3.0
contourpy==1.3.3
coverage==7.11.3
cycler==0.12.1
dill==0.4.0
Django==5.2.8
-e git+https://github.com/arunakulat/dutchbay-epc-model.git@2fb6ea38bae8f9dd0aefa442f64b41c174f7c62e#egg=dutchbay_v13
fastapi==0.121.2
flake8==7.3.0
fonttools==4.60.1
hypothesis==6.147.0
idna==3.11
iniconfig==2.3.0
isort==7.0.0
kiwisolver==1.4.9
linecache2==1.0.0
markdown-it-py==4.0.0
matplotlib==3.10.7
mccabe==0.7.0
mdurl==0.1.2
mock==5.2.0
mypy==1.18.2
mypy_extensions==1.1.0
numpy==2.3.4
numpy-financial==1.0.0
packaging==25.0
pandas==2.3.3
pathspec==0.12.1
pillow==12.0.0
platformdirs==4.5.0
pluggy==1.6.0
pycodestyle==2.14.0
pydantic==2.12.4
pydantic_core==2.41.5
pyflakes==3.4.0
Pygments==2.19.2
pylint==4.0.3
pyparsing==3.2.5
pytest==9.0.1
pytest-cov==7.0.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.2.1
pytokens==0.3.0
pytz==2025.2
PyYAML==6.0.3
rich==14.2.0
ruff==0.14.4
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.5.3
starlette==0.49.3
stevedore==5.5.0
tomlkit==0.13.3
tqdm==4.67.1
traceback2==1.4.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
unittest2==1.1.0
